Add files via upload
Here is the example of Web Development i.e JavaScript (JS) and JQuery. In this small code of JavaScript, we have used the click function which changes the images on click and the UI looks beautiful. This block of code can also be executed in switching the mode just like we switch light mode to dark mode with same concept.